Textract/requirements.txt
langgraph
langchain
psycopg2-binary
pgvector
google-generativeai
tavily-python
python-dotenv
requests
pydantic
uvicorn
fastapi
tqdm
PyPDF2
python-multipart
sentence-transformers

docx
exceptions
